pyproject.toml
[tool.poetry]
name = "proxydetect"
version = "0.1.0"
description = ""
authors = ["miqsoft <mksaxhandball@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
jinja2 = "^3.1.4"
click = "^8.1.7"
python-vagrant = "^1.0.0"
python-digitalocean = "^1.17.0"
python-dotenv = "^1.0.1"
pydo = "^0.8.0"
parallel-ssh = "^2.13.0"
aioquic = "^1.2.0"
pyyaml = "^6.0.2"


[tool.poetry.scripts]
main = "proxydetect:main3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
